Remove `var_key` from jobs in favor of the `variables` key.
The entire concept of `var_key`...works, but it's one of those things that you hope people don't look too closely at so you don't have to answer super embarrassing questions as to why it exists and why you couldn't come up with something better.  Instead of providing a random key to arbitrarily lock the job template's variables instead provide a mechanism to supply key/value pairs to the jobs as variables and then use one of them to lookup the appropriate value in the template.